+++ apps/ae_socket_connector/lib/client_runner.ex
@@ -5,9 +5,11 @@ defmodule ClientRunner do
   @ae_url "ws://localhost:3014/channel"
   @network_id "my_test"
 
+
   # @ae_url "wss://testnet.demo.aeternity.io/channel"
   # @network_id "ae_uat"
 
+  import ClientRunnerHelper
   # TODO :local always? produces a cast, could we move this to :local runner
 
   defmacro ae_url, do: @ae_url
@@ -27,14 +29,14 @@ defmodule ClientRunner do
 
   def joblist(),
     do: [
-      # &hello_fsm_v3/3,
-      # &hello_fsm_v2/3,
-      # &withdraw_after_reconnect_v2/3,
-      # # # # # &withdraw_after_reestablish/3,
-      # &backchannel_jobs_v2/3,
-      # &close_solo_v2/3,
-      # &close_mutual_v2/3,
-      # &reconnect_jobs_v2/3,
+      &hello_fsm_v3/3,
+      &hello_fsm_v2/3,
+      &withdraw_after_reconnect_v2/3,
+      # &withdraw_after_reestablish/3,
+      &backchannel_jobs_v2/3,
+      &close_solo_v2/3,
+      &close_mutual_v2/3,
+      &reconnect_jobs_v2/3,
       &contract_jobs_v2/3
       # &reestablish_jobs/3,
       # &query_after_reconnect/3,
@@ -54,10 +56,6 @@ defmodule ClientRunner do
   # %{
   # {:initiator, %{message: {:channels_update, 1, :transient, "channels.update"}, next: {:run_job}, fuzzy: 3}},
   # }
-
-  def resume_runner(client_runner) do
-    GenServer.cast(client_runner, {:match_jobs, {}})
-  end
 
   def hello_fsm_v2({initiator, _intiator_account}, {responder, _responder_account}, runner_pid),
     do: [
@@ -259,7 +257,7 @@ defmodule ClientRunner do
        }}
     ]
 
-  def connection_callback_in_details(callback_pid, color) do
+  def connection_callback(callback_pid, color) do
     %SocketConnector.ConnectionCallbacks{
       sign_approve: fn round_initiator, round, auto_approval, human ->
         Logger.debug(
@@ -295,60 +293,6 @@ defmodule ClientRunner do
     }
   end
 
-  def connection_callback(callback_pid, color) do
-    %SocketConnector.ConnectionCallbacks{
-      # auto approval is the suggested response. Typically if initiated by us and matches with our request we should approve.
-      sign_approve: fn round_initiator, round, auto_approval, human ->
-        Logger.debug(
-          "Sign request for round: #{inspect(round)}, initated by: #{inspect(round_initiator)}. auto_approval: #{
-            inspect(auto_approval)
-          }, containing: #{inspect(human)}",
-          ansi_color: color
-        )
-
-        auto_approval
-      end,
-      channels_info: fn round_initiator, round, method ->
-        Logger.debug(
-          "info callback received round is: #{inspect(round)} round_initiator is: #{inspect(round_initiator)} method is #{
-            inspect(method)
-          }}",
-          ansi_color: color
-        )
-
-        # &teardown_on_channel_creation
-        # case method do
-        #   "funding_signed" ->
-        #     GenServer.cast(callback_pid, {:process_job_lists})
-        #   _ -> :ok
-        # end
-      end,
-      channels_update: fn round_initiator, round, method ->
-        Logger.debug(
-          "callback received round is: #{inspect(round)} round_initiator is: #{inspect(round_initiator)} method is #{
-            inspect(method)
-          }}",
-          ansi_color: color
-        )
-
-        case round_initiator do
-          :self ->
-            GenServer.cast(callback_pid, {:process_job_lists})
-
-          :other ->
-            GenServer.cast(callback_pid, {:process_job_lists})
-
-          :transient ->
-            # connect/reconnect allow grace period before resuming
-            spawn(fn ->
-              Process.sleep(500)
-              GenServer.cast(callback_pid, {:process_job_lists})
-            end)
-        end
-      end
-    }
-  end
-
   def seperate_jobs(job_list) do
     {filter_jobs(job_list, :initiator), filter_jobs(job_list, :responder)}
   end
@@ -371,7 +315,7 @@ defmodule ClientRunner do
           priv_key: priv_key,
           session: state_channel_configuration,
           role: role,
-          connection_callbacks: connection_callback_in_details(self(), color)
+          connection_callbacks: connection_callback(self(), color)
           # connection_callbacks: connection_callback(self(), color)
         },
         ae_url: ae_url,
@@ -396,29 +340,6 @@ defmodule ClientRunner do
        #  socket_holder_name: name,
        color: [ansi_color: color]
      }}
-  end
-
-  # TODO time to move this module to ex unit
-  def expected(response, {account1, value1}, {account2, value2}) do
-    expect =
-      Enum.sort([
-        %{"account" => account1, "balance" => value1},
-        %{"account" => account2, "balance" => value2}
-      ])
-
-    Logger.debug("expected #{inspect(Enum.sort(expect))}")
-    Logger.debug("got      #{inspect(response)}")
-    ^expect = Enum.sort(response)
-  end
-
-  def assert_funds_job({account1, value1}, {account2, value2}) do
-    {:sync,
-     fn pid, from ->
-       SocketConnector.query_funds(pid, from)
-     end,
-     fn result ->
-       expected(result, {account1, value1}, {account2, value2})
-     end}
   end
 
   def run_next(match) do
@@ -506,149 +427,6 @@ defmodule ClientRunner do
     end
 
     {:noreply, state}
-  end
-
-  # def handle_cast({:process_job_lists}, state) do
-  #   remaining_jobs =
-  #     case Enum.count(state.job_list) do
-  #       0 ->
-  #         Logger.debug("End of sequence", state.color)
-  #         []
-
-  #       remaining ->
-  #         Logger.debug("Sequence remainin jobs #{inspect(remaining)}", state.color)
-  #         [{mode, fun, assert_fun} | rest] = state.job_list
-
-  #         case mode do
-  #           :async ->
-  #             SessionHolder.run_action(state.pid_session_holder, fun)
-
-  #           :sync ->
-  #             response = SessionHolder.run_action_sync(state.pid_session_holder, fun)
-
-  #             case assert_fun do
-  #               :empty -> :empty
-  #               _ -> assert_fun.(response)
-  #             end
-
-  #             Logger.debug("sync response is: #{inspect(response)}", state.color)
-  #             GenServer.cast(self(), {:process_job_lists})
-
-  #           :local ->
-  #             fun.(self(), state.pid_session_holder)
-  #         end
-
-  #         rest
-  #     end
-
-  #   {:noreply, %__MODULE__{state | job_list: remaining_jobs}}
-  # end
-
-  def empty_jobs(interval) do
-    Enum.map(interval, fn count ->
-      {:local,
-       fn _client_runner, _pid_session_holder ->
-         Logger.debug("doing nothing #{inspect(count)}", ansi_color: :white)
-       end, :empty}
-    end)
-  end
-
-  # def hello_fsm({initiator, _intiator_account}, {responder, _responder_account}, runner_pid) do
-  #   jobs_initiator = [
-  #     {:async, fn pid -> SocketConnector.leave(pid) end, :empty},
-  #     sequence_finish_job(runner_pid, initiator)
-  #   ]
-
-  #   {jobs_initiator, [sequence_finish_job(runner_pid, responder)]}
-  # end
-
-  def contract_jobs({initiator, intiator_account}, {responder, responder_account}, runner_pid) do
-    initiator_contract = {TestAccounts.initiatorPubkeyEncoded(), "../../contracts/TicTacToe.aes"}
-    # correct path if started in shell...
-    # initiator_contract = {TestAccounts.initiatorPubkeyEncoded(), "contracts/TicTacToe.aes"}
-    # responder_contract = {TestAccounts.responderPubkeyEncoded(), "contracts/TicTacToe.aes"}
-
-    jobs_initiator = [
-      {:async, fn pid -> SocketConnector.initiate_transfer(pid, 2) end, :empty},
-      assert_funds_job(
-        {intiator_account, 6_999_999_999_997},
-        {responder_account, 4_000_000_000_003}
-      ),
-      {:async, fn pid -> SocketConnector.new_contract(pid, initiator_contract) end, :empty},
-      {:async,
-       fn pid ->
-         SocketConnector.call_contract(
-           pid,
-           initiator_contract,
-           'make_move',
-           ['11', '1']
-         )
-       end, :empty},
-      {:sync,
-       fn pid, from ->
-         SocketConnector.get_contract_reponse(
-           pid,
-           initiator_contract,
-           'make_move',
-           from
-         )
-       end, :empty},
-      {:async, fn pid -> SocketConnector.initiate_transfer(pid, 3) end, :empty},
-      # hard coded to put 10 coins in the created contract
-      assert_funds_job(
-        {intiator_account, 6_999_999_999_984},
-        {responder_account, 4_000_000_000_006}
-      ),
-      {:async,
-       fn pid ->
-         SocketConnector.withdraw(pid, 1_000_000)
-       end, :empty},
-      assert_funds_job(
-        {intiator_account, 6_999_998_999_984},
-        {responder_account, 4_000_000_000_006}
-      ),
-      {:async, fn pid -> SocketConnector.initiate_transfer(pid, 4) end, :empty},
-      assert_funds_job(
-        {intiator_account, 6_999_998_999_980},
-        {responder_account, 4_000_000_000_010}
-      ),
-      {:async, fn pid -> SocketConnector.initiate_transfer(pid, 5) end, :empty},
-      {:async,
-       fn pid ->
-         SocketConnector.deposit(pid, 500_000)
-       end, :empty},
-      assert_funds_job(
-        {intiator_account, 6_999_999_499_975},
-        {responder_account, 4_000_000_000_015}
-      ),
-      sequence_finish_job(runner_pid, initiator)
-    ]
-
-    jobs_responder =
-      empty_jobs(1..8) ++
-        [
-          {:async,
-           fn pid ->
-             SocketConnector.call_contract(
-               pid,
-               initiator_contract,
-               'make_move',
-               ['11', '2']
-             )
-           end, :empty},
-          {:sync,
-           fn pid, from ->
-             SocketConnector.get_contract_reponse(
-               pid,
-               initiator_contract,
-               'make_move',
-               from
-             )
-           end, :empty},
-          sequence_finish_job(runner_pid, responder)
-        ]
-
-    {jobs_initiator, jobs_responder}
   end
 
   def contract_jobs_v2({initiator, intiator_account}, {responder, responder_account}, runner_pid) do
@@ -801,195 +579,6 @@ defmodule ClientRunner do
     ]
   end
 
-  # query after violent reestablish
-  def query_after_reconnect({initiator, intiator_account}, {responder, responder_account}, runner_pid) do
-    jobs_initiator =
-      empty_jobs(1..1) ++
-        [
-          assert_funds_job(
-            {intiator_account, 7_000_000_000_001},
-            {responder_account, 3_999_999_999_999}
-          ),
-          {:local,
-           fn client_runner, pid_session_holder ->
-             Logger.debug("killing previous connection 1")
-             SessionHolder.kill_connection(pid_session_holder)
-             Logger.debug("reestablish 1")
-             SessionHolder.reestablish(pid_session_holder)
-             GenServer.cast(client_runner, {:process_job_lists})
-           end, :empty},
-          assert_funds_job(
-            {intiator_account, 7_000_000_000_001},
-            {responder_account, 3_999_999_999_999}
-          ),
-          sequence_finish_job(runner_pid, responder)
-        ]
-
-    jobs_responder = [
-      assert_funds_job(
-        {intiator_account, 6_999_999_999_999},
-        {responder_account, 4_000_000_000_001}
-      ),
-      {:async, fn pid -> SocketConnector.initiate_transfer(pid, 2) end, :empty},
-      pause_job(1000),
-      assert_funds_job(
-        {intiator_account, 7_000_000_000_001},
-        {responder_account, 3_999_999_999_999}
-      ),
-      sequence_finish_job(runner_pid, initiator)
-    ]
-
-    {jobs_initiator, jobs_responder}
-  end
-
-  def teardown_on_channel_creation({initiator, _intiator_account}, {responder, _responder_account}, runner_pid) do
-    # empty_jobs(1..1) ++
-    jobs_initiator = [
-      {:local,
-       fn client_runner, pid_session_holder ->
-         Logger.debug("close")
-         # SessionHolder.kill_connection(pid_session_holder)
-         SessionHolder.close_connection(pid_session_holder)
-         GenServer.cast(client_runner, {:process_job_lists})
-       end, :empty},
-      pause_job(10000),
-      {:local,
-       fn client_runner, pid_session_holder ->
-         Logger.debug("reestablish 1")
-         SessionHolder.reestablish(pid_session_holder)
-         GenServer.cast(client_runner, {:process_job_lists})
-       end, :empty},
-      # assert_funds_job(
-      #   {intiator_account, 6_999_999_999_997},
-      #   {responder_account, 4_000_000_000_003}
-      # ),
-      {:sync,
-       fn pid, from ->
-         SocketConnector.query_funds(pid, from)
-       end, :empty},
-      pause_job(5000),
-      sequence_finish_job(runner_pid, responder)
-    ]
-
-    jobs_responder = [
-      sequence_finish_job(runner_pid, initiator)
-    ]
-
-    {jobs_initiator, jobs_responder}
-  end
-
-  # currently broken
-  def reestablish_jobs({initiator, intiator_account}, {responder, responder_account}, runner_pid) do
-    jobs_initiator = [
-      {:async, fn pid -> SocketConnector.initiate_transfer(pid, 2) end, :empty},
-      {:sync,
-       fn pid, from ->
-         SocketConnector.query_funds(pid, from)
-       end, :empty},
-      {:async, fn pid -> SocketConnector.leave(pid) end, :empty},
-      {:local,
-       fn client_runner, pid_session_holder ->
-         SessionHolder.reestablish(pid_session_holder)
-         GenServer.cast(client_runner, {:process_job_lists})
-       end, :empty},
-      sequence_finish_job(runner_pid, initiator)
-    ]
-
-    jobs_responder =
-      empty_jobs(1..2) ++
-        [
-          {:local,
-           fn client_runner, pid_session_holder ->
-             Logger.debug("reestablish 1")
-             SessionHolder.reestablish(pid_session_holder)
-             GenServer.cast(client_runner, {:process_job_lists})
-           end, :empty},
-          pause_job(5000),
-          assert_funds_job(
-            {intiator_account, 6_999_999_999_997},
-            {responder_account, 4_000_000_000_003}
-          ),
-          pause_job(5000),
-          # reestablish without leave
-          {:async, fn pid -> SocketConnector.leave(pid) end, :empty},
-          pause_job(5000),
-          {:local,
-           fn client_runner, pid_session_holder ->
-             Logger.debug("reestablish 2")
-             SessionHolder.reestablish(pid_session_holder)
-             GenServer.cast(client_runner, {:process_job_lists})
-           end, :empty},
-          pause_job(5000),
-          assert_funds_job(
-            {intiator_account, 6_999_999_999_997},
-            {responder_account, 4_000_000_000_003}
-          ),
-          # pause_job(5000),
-          sequence_finish_job(runner_pid, responder)
-        ]
-
-    {jobs_initiator, jobs_responder}
-  end
-
-  def pause_job(delay) do
-    {:local,
-     fn client_runner, _pid_session_holder ->
-       Logger.debug("requested pause for: #{inspect(delay)}ms")
-
-       spawn(fn ->
-         Process.sleep(delay)
-         resume_runner(client_runner)
-       end)
-     end, :empty}
-  end
-
-  # def reconnect_jobs({initiator, intiator_account}, {responder, responder_account}, runner_pid) do
-  #   jobs_initiator = [
-  #     assert_funds_job(
-  #       {intiator_account, 6_999_999_999_999},
-  #       {responder_account, 4_000_000_000_001}
-  #     ),
-  #     {:async, fn pid -> SocketConnector.initiate_transfer(pid, 2) end, :empty},
-  #     assert_funds_job(
-  #       {intiator_account, 6_999_999_999_997},
-  #       {responder_account, 4_000_000_000_003}
-  #     ),
-  #     sequence_finish_job(runner_pid, initiator)
-  #   ]
-
-  #   jobs_responder =
-  #     empty_jobs(1..1) ++
-  #       [
-  #         {:sync,
-  #          fn pid, from ->
-  #            SocketConnector.query_funds(pid, from)
-  #          end, :empty},
-  #         {:local,
-  #          fn client_runner, pid_session_holder ->
-  #            SessionHolder.close_connection(pid_session_holder)
-  #            GenServer.cast(client_runner, {:process_job_lists})
-  #          end, :empty},
-  #         pause_job(3000),
-  #         {:local,
-  #          fn client_runner, pid_session_holder ->
-  #            SessionHolder.reconnect(pid_session_holder)
-  #            GenServer.cast(client_runner, {:process_job_lists})
-  #          end, :empty},
-  #         assert_funds_job(
-  #           {intiator_account, 6_999_999_999_997},
-  #           {responder_account, 4_000_000_000_003}
-  #         ),
-  #         {:async, fn pid -> SocketConnector.initiate_transfer(pid, 2) end, :empty},
-  #         assert_funds_job(
-  #           {intiator_account, 6_999_999_999_999},
-  #           {responder_account, 4_000_000_000_001}
-  #         ),
-  #         sequence_finish_job(runner_pid, responder)
-  #       ]
-
-  #   {jobs_initiator, jobs_responder}
-  # end
-
   def reconnect_jobs_v2({initiator, intiator_account}, {responder, responder_account}, runner_pid),
     do: [
       {:initiator,
@@ -1115,20 +704,6 @@ defmodule ClientRunner do
     {jobs_initiator, jobs_responder}
   end
 
-  # def close_solo({initiator, _intiator_account}, {responder, _responder_account}, runner_pid) do
-  #   jobs_initiator = [
-  #     {:async, fn pid -> SocketConnector.initiate_transfer(pid, 5) end, :empty},
-  #     close_solo_job(),
-  #     sequence_finish_job(runner_pid, initiator)
-  #   ]
-
-  #   jobs_responder = [
-  #     sequence_finish_job(runner_pid, responder)
-  #   ]
-
-  #   {jobs_initiator, jobs_responder}
-  # end
-
   def close_solo_v2({initiator, _intiator_account}, {responder, _responder_account}, runner_pid),
     do: [
       {:initiator,
@@ -1156,21 +731,6 @@ defmodule ClientRunner do
          next: sequence_finish_job(runner_pid, responder)
        }}
     ]
-
-  # def close_mutual({initiator, _intiator_account}, {responder, _responder_account}, runner_pid) do
-  #   jobs_initiator = [
-  #     {:async, fn pid -> SocketConnector.initiate_transfer(pid, 5) end, :empty},
-  #     {:sync, fn pid, from -> SocketConnector.get_poi(pid, from) end, :empty},
-  #     close_mutual_job(),
-  #     sequence_finish_job(runner_pid, initiator)
-  #   ]
-
-  #   jobs_responder = [
-  #     sequence_finish_job(runner_pid, responder)
-  #   ]
-
-  #   {jobs_initiator, jobs_responder}
-  # end
 
   def close_mutual_v2({initiator, _intiator_account}, {responder, _responder_account}, runner_pid),
     do: [
@@ -1206,140 +766,6 @@ defmodule ClientRunner do
          next: sequence_finish_job(runner_pid, responder)
        }}
     ]
-
-  def sequence_finish_job(runner_pid, name) do
-    {:local,
-     fn _client_runner, _pid_session_holder ->
-       Logger.info("Sending termination for #{inspect(name)}")
-       send(runner_pid, {:test_finished, name})
-     end, :empty}
-  end
-
-  # reconstruct https://www.pivotaltracker.com/n/projects/2124891/stories/167944617
-  def withdraw_after_reestablish(
-        {initiator, _intiator_account},
-        {responder, _responder_account},
-        runner_pid
-      ) do
-    jobs_initiator = [
-      {:async, fn pid -> SocketConnector.leave(pid) end, :empty},
-      {:local, fn _client_runner, pid_session_holder -> SessionHolder.reestablish(pid_session_holder) end, :empty},
-      {:async,
-       fn pid ->
-         SocketConnector.withdraw(pid, 1_000_000)
-       end, :empty},
-      {:async,
-       fn pid ->
-         SocketConnector.deposit(pid, 1_200_000)
-       end, :empty},
-      sequence_finish_job(runner_pid, initiator)
-    ]
-
-    jobs_responder =
-      empty_jobs(1..1) ++
-        [
-          {:local,
-           fn _client_runner, pid_session_holder ->
-             SessionHolder.reestablish(pid_session_holder)
-           end, :empty},
-          sequence_finish_job(runner_pid, responder)
-        ]
-
-    {jobs_initiator, jobs_responder}
-  end
-
-  # reconstruct https://www.pivotaltracker.com/n/projects/2124891/stories/167944617
-  def withdraw_after_reconnect(
-        {initiator, _intiator_account},
-        {responder, _responder_account},
-        runner_pid
-      ) do
-    jobs_initiator = [
-      {:local,
-       fn client_runner, pid_session_holder ->
-         SessionHolder.close_connection(pid_session_holder)
-         GenServer.cast(client_runner, {:process_job_lists})
-       end, :empty},
-      {:local,
-       fn client_runner, _pid_session_holder ->
-         Process.sleep(1000)
-         GenServer.cast(client_runner, {:process_job_lists})
-       end, :empty},
-      {:local,
-       fn client_runner, pid_session_holder ->
-         SessionHolder.reconnect(pid_session_holder)
-         GenServer.cast(client_runner, {:process_job_lists})
-       end, :empty},
-      {:async,
-       fn pid ->
-         SocketConnector.withdraw(pid, 1_000_000)
-       end, :empty},
-      {:async,
-       fn pid ->
-         SocketConnector.deposit(pid, 1_200_000)
-       end, :empty},
-      sequence_finish_job(runner_pid, initiator)
-    ]
-
-    jobs_responder = [
-      sequence_finish_job(runner_pid, responder)
-    ]
-
-    {jobs_initiator, jobs_responder}
-  end
-
-  # https://github.com/aeternity/protocol/blob/master/node/api/channels_api_usage.md#example
-  # def backchannel_jobs({initiator, intiator_account}, {responder, responder_account}, runner_pid) do
-  #   jobs_initiator = [
-  #     {:local,
-  #      fn client_runner, pid_session_holder ->
-  #        SessionHolder.close_connection(pid_session_holder)
-  #        GenServer.cast(client_runner, {:process_job_lists})
-  #      end, :empty},
-  #     pause_job(10000),
-  #     {:local,
-  #      fn client_runner, pid_session_holder ->
-  #        SessionHolder.reconnect(pid_session_holder)
-  #        GenServer.cast(client_runner, {:process_job_lists})
-  #      end, :empty},
-  #     assert_funds_job(
-  #       {intiator_account, 7_000_000_000_003},
-  #       {responder_account, 3_999_999_999_997}
-  #     ),
-  #     pause_job(5000),
-  #     {:async, fn pid -> SocketConnector.initiate_transfer(pid, 5) end, :empty},
-  #     assert_funds_job(
-  #       {intiator_account, 6_999_999_999_998},
-  #       {responder_account, 4_000_000_000_002}
-  #     ),
-  #     sequence_finish_job(runner_pid, initiator)
-  #   ]
-
-  #   jobs_responder = [
-  #     pause_job(3000),
-  #     assert_funds_job(
-  #       {intiator_account, 6_999_999_999_999},
-  #       {responder_account, 4_000_000_000_001}
-  #     ),
-  #     {:async, fn pid -> SocketConnector.initiate_transfer(pid, 2) end, :empty},
-  #     {:async, fn pid -> SocketConnector.initiate_transfer(pid, 3) end, :empty},
-  #     {:async,
-  #      fn pid ->
-  #        SocketConnector.initiate_transfer(pid, 4, fn to_sign ->
-  #          SessionHolder.backchannel_sign_request(initiator, to_sign)
-  #          # GenServer.call(initiator, {:sign_request, to_sign})
-  #        end)
-  #      end, :empty},
-  #     assert_funds_job(
-  #       {intiator_account, 7_000_000_000_003},
-  #       {responder_account, 3_999_999_999_997}
-  #     ),
-  #     {:async, fn pid -> SocketConnector.initiate_transfer(pid, 5) end, :empty},
-  #     sequence_finish_job(runner_pid, responder)
-  #   ]
-
-  #   {jobs_initiator, jobs_responder}
-  # end
 
   def gen_name(name, suffix) do
     String.to_atom(to_string(name) <> Integer.to_string(suffix))
